#713: vrt coll: fix compile issues with hash for gcc-5
